add usersocial profile hook
