+++ src/components/UserProfileSettings/UserProfileSocialSettings.tsx
@@ -4,24 +4,17 @@ import { css } from 'styled-components';
 import { useSelector } from 'react-redux';
 import { useNavigate } from 'react-router-dom';
 
-//Hooks
-import { useDisclosure } from 'common';
-import { useAccount } from 'hooks';
-import { useGetSocialsStatus } from 'queries';
-
 // Helpers
-import { generateVerificationProof } from 'modules/rewards/utils/generateVerificationProof';
-import { appConfig } from 'config';
-import { walletToCAIP10 } from 'helpers/w2w';
 import UnlockProfileWrapper, { UNLOCK_PROFILE_TYPE } from 'components/chat/unlockProfile/UnlockProfileWrapper';
 import APP_PATHS from 'config/AppPaths';
 
 //Components
-import { Box, DiscordProfile, EmailProfile, TelegramProfile, Text } from 'blocks';
+import { Box, Text } from 'blocks';
 import { AddEmail } from './AddEmail';
 import AddTelegram from './AddTelegram';
 import AddDiscord from './AddDiscord';
 import UserProfileSettingsItem from './UserProfileSettingsItem';
+import { useSocialHandles } from 'modules/dashboard/hooks/useSocialHandles';
 
 type UserProfileSocialSettingsType = {
   errorMessage?: string;
@@ -30,60 +23,27 @@ type UserProfileSocialSettingsType = {
   setSuccessMessage: (successMessage: string) => void;
 };
 
-type SocialHandleStatusType = {
-  email: string | null;
-  discord_username: string | null;
-  telegram_username: string | null;
-};
-
 const UserProfileSocialSettings: FC<UserProfileSocialSettingsType> = ({ setErrorMessage, setSuccessMessage }) => {
-  const modalControl = useDisclosure();
-  const telegramModalControl = useDisclosure();
-  const discordModalControl = useDisclosure();
-  const { account } = useAccount();
   const navigate = useNavigate();
-
-  const [socialHandleStatus, setSocialHandleStatus] = useState<SocialHandleStatusType>();
   const [isAuthModalVisible, setIsAuthModalVisible] = useState(true);
 
   const { userPushSDKInstance } = useSelector((state: any) => {
     return state.user;
   });
 
-  // Getting user Id by wallet address
-  const channelAddress = walletToCAIP10({ account });
-
-  const { mutate: fetchSocialStatus, isPending } = useGetSocialsStatus();
-
-  // Fetch social status logic
-  const getSocialStatus = async () => {
-    if (!channelAddress) return;
-
-    const data = {};
-
-    const verificationProof = await generateVerificationProof(data, userPushSDKInstance);
-
-    if (!verificationProof) return;
-
-    if (channelAddress) {
-      fetchSocialStatus(
-        { channelAddress, verificationProof: verificationProof as string },
-        {
-          onError: (error) => {
-            setErrorMessage('Failed to fetch social status.');
-            console.error('Error fetching social status:', error);
-          },
-          onSuccess: (data) => {
-            setSocialHandleStatus(data);
-          },
-        }
-      );
-    }
-  };
+  const {
+    socialHandlesList,
+    modalControl,
+    telegramModalControl,
+    discordModalControl,
+    isPending,
+    fetchStatus,
+    channelAddress,
+  } = useSocialHandles(setErrorMessage, true, userPushSDKInstance);
 
   useEffect(() => {
     if (!userPushSDKInstance || !channelAddress) return;
-    getSocialStatus();
+    fetchStatus();
   }, [userPushSDKInstance, channelAddress]);
 
   useEffect(() => {
@@ -94,30 +54,6 @@ const UserProfileSocialSettings: FC<UserProfileSocialSettingsType> = ({ setError
     setIsAuthModalVisible(false);
     navigate(APP_PATHS.WelcomeDashboard);
   };
-
-  const itemList = [
-    {
-      icon: () => <EmailProfile height={23} />,
-      itemTitle: 'Email',
-      itemDescription: 'Receive notifications in your email inbox',
-      onClick: () => modalControl.open(),
-      userStatus: socialHandleStatus?.email || null,
-    },
-    appConfig?.telegramExternalURL && {
-      icon: () => <TelegramProfile height={23} />,
-      itemTitle: 'Telegram',
-      itemDescription: 'Receive notifications as Telegram messages',
-      onClick: () => telegramModalControl.open(),
-      userStatus: socialHandleStatus?.telegram_username || null,
-    },
-    appConfig?.discordExternalURL && {
-      icon: () => <DiscordProfile height={23} />,
-      itemTitle: 'Discord',
-      itemDescription: 'Receive notifications as Discord messages',
-      onClick: () => discordModalControl.open(),
-      userStatus: socialHandleStatus?.discord_username || null,
-    },
-  ].filter(Boolean);
 
   return (
     <Box>
@@ -138,7 +74,7 @@ const UserProfileSocialSettings: FC<UserProfileSocialSettingsType> = ({ setError
       </Box>
 
       <Box padding="spacing-sm spacing-none spacing-none spacing-none">
-        {itemList?.map((item) => (
+        {socialHandlesList?.map((item) => (
           <UserProfileSettingsItem
             item={item}
             isPending={isPending}
@@ -168,7 +104,7 @@ const UserProfileSocialSettings: FC<UserProfileSocialSettingsType> = ({ setError
       {modalControl.isOpen && (
         <AddEmail
           modalControl={modalControl}
-          refetchSocialHandleStatus={getSocialStatus}
+          refetchSocialHandleStatus={fetchStatus}
           setErrorMessage={setErrorMessage}
           setSuccessMessage={setSuccessMessage}
         />
@@ -177,7 +113,7 @@ const UserProfileSocialSettings: FC<UserProfileSocialSettingsType> = ({ setError
       {telegramModalControl.isOpen && (
         <AddTelegram
           modalControl={telegramModalControl}
-          refetchSocialHandleStatus={getSocialStatus}
+          refetchSocialHandleStatus={fetchStatus}
           setErrorMessage={setErrorMessage}
           setSuccessMessage={setSuccessMessage}
         />
@@ -186,7 +122,7 @@ const UserProfileSocialSettings: FC<UserProfileSocialSettingsType> = ({ setError
       {discordModalControl.isOpen && (
         <AddDiscord
           modalControl={discordModalControl}
-          refetchSocialHandleStatus={getSocialStatus}
+          refetchSocialHandleStatus={fetchStatus}
           setErrorMessage={setErrorMessage}
           setSuccessMessage={setSuccessMessage}
         />
