+++ src/config/config-dev.js
@@ -83,7 +83,7 @@ export const config = {
   },
 
   // social media integration
-  telegramExternalURL: 'https://t.me/PushStagingBot',
+  telegramExternalURL: 'https://t.me/PushCommDevBot',
   discordExternalURL: 'https://discordapp.com/users/pushlinkbot',
 };
